Rename subprogram to `buffer-advance`
