+++ src/app.rs
@@ -22,7 +22,7 @@ use crate::service;
 #[cfg(feature = "api")]
 use crate::{api, internal_events::ApiStarted};
 use crate::{
-    buffer_fix,
+    buffer_advance,
     cli::{handle_config_errors, Color, LogFormat, Opts, RootOpts, SubCommand},
     config::{self},
     generate, graph, heartbeat, list,
@@ -145,7 +145,7 @@ impl Application {
 
                 if let Some(s) = sub_command {
                     let code = match s {
-                        SubCommand::BufferFix(b) => buffer_fix::cmd(&b).await,
+                        SubCommand::BufferAdvance(b) => buffer_advance::cmd(&b).await,
                         SubCommand::Generate(g) => generate::cmd(&g),
                         SubCommand::Graph(g) => graph::cmd(&g),
                         SubCommand::Config(c) => config::cmd(&c),
