+++ src/cli.rs
@@ -8,7 +8,7 @@ use crate::service;
 use crate::tap;
 #[cfg(feature = "api-client")]
 use crate::top;
-use crate::{buffer_fix, config, generate, get_version, graph, list, unit_test, validate};
+use crate::{buffer_advance, config, generate, get_version, graph, list, unit_test, validate};
 
 #[derive(Parser, Debug)]
 #[clap(rename_all = "kebab-case")]
@@ -210,7 +210,7 @@ pub enum SubCommand {
     Vrl(vrl_cli::Opts),
 
     /// Manipulate the buffer ledger.
-    BufferFix(buffer_fix::Opts),
+    BufferAdvance(buffer_advance::Opts),
 }
 
 #[derive(Debug, Clone, PartialEq)]
