+++ src/lib.rs
@@ -26,7 +26,7 @@ extern crate derivative;
 #[global_allocator]
 static ALLOC: tikv_jemallocator::Jemalloc = tikv_jemallocator::Jemalloc;
 
-mod buffer_fix;
+mod buffer_advance;
 #[macro_use]
 #[allow(unreachable_pub)]
 pub mod config;
